Move client implementation to pkg/proxy
