+++ cmd/npv/app/dump.go
@@ -7,6 +7,8 @@ import (
 	"io"
 
 	"github.com/spf13/cobra"
+
+	"github.com/cybozu-go/network-policy-viewer/pkg/proxy"
 )
 
 func init() {
@@ -31,12 +33,12 @@ func runDump(ctx context.Context, stdout, stderr io.Writer, name string) error {
 		return err
 	}
 
-	client, err := createCiliumClient(ctx, stderr, clientset, dynamicClient, rootOptions.namespace, name)
+	client, err := proxy.CreateCiliumClient(ctx, stderr, clientset, dynamicClient, rootOptions.namespace, name)
 	if err != nil {
 		return err
 	}
 
-	data, err := client.dumpEndpoint(ctx, rootOptions.namespace, name)
+	data, err := client.DumpEndpoint(ctx, rootOptions.namespace, name)
 	if err != nil {
 		return err
 	}
