+++ cmd/npv/app/inspect.go
@@ -16,6 +16,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 
 	"github.com/cybozu-go/network-policy-viewer/pkg/cidr"
+	"github.com/cybozu-go/network-policy-viewer/pkg/proxy"
 )
 
 var inspectOptions struct {
@@ -126,7 +127,7 @@ func parseInspectOptions() {
 }
 
 func runInspectOnPod(ctx context.Context, stderr io.Writer, clientset *kubernetes.Clientset, dynamicClient *dynamic.DynamicClient, filter policyFilter, pod *corev1.Pod) ([]inspectEntry, error) {
-	client, err := createCiliumClient(ctx, stderr, clientset, dynamicClient, pod.Namespace, pod.Name)
+	client, err := proxy.CreateCiliumClient(ctx, stderr, clientset, dynamicClient, pod.Namespace, pod.Name)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create Cilium client: %w", err)
 	}
@@ -136,7 +137,7 @@ func runInspectOnPod(ctx context.Context, stderr io.Writer, clientset *kubernete
 		return nil, err
 	}
 
-	policies, err := client.queryPolicyMap(ctx, pod.Namespace, pod.Name)
+	policies, err := client.QueryPolicyMap(ctx, pod.Namespace, pod.Name)
 	if err != nil {
 		return nil, err
 	}
@@ -182,7 +183,7 @@ func runInspectOnPod(ctx context.Context, stderr io.Writer, clientset *kubernete
 			if idObj.IsReservedIdentity() {
 				entry.Example = "reserved:" + idObj.String()
 			} else if idObj.HasLocalScope() {
-				c, err := client.getCIDRForIdentity(ctx, p.Key.Identity)
+				c, err := client.GetCIDRForIdentity(ctx, p.Key.Identity)
 				if err != nil {
 					return nil, err
 				}
